// src/app/core/interfaces/global/management-topic.interface.ts
export interface ManagementTopic {
  _id: string;
  assignedTopic: string;
  enrollment: string;
  statusManagementTopic: boolean;
}

// import { Enrollment } from "../enrollment.interface";

// export interface ManagementTopic {
//   _id: string;
//   assignedTopic: string;
//   enrollment: Enrollment; 
//   statusManagementTopic: boolean;
// }